% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_peptides.R
\name{AddUniquenessColumn}
\alias{AddUniquenessColumn}
\title{AddUniquenessColumn}
\usage{
AddUniquenessColumn(
  df,
  peptide_col = "Precursor.Id",
  protein_col = "Protein.Group"
)
}
\arguments{
\item{df}{Data frame containing the data}

\item{peptide_col}{Name of the peptide/precursor column}

\item{protein_col}{Name of the protein column}
}
\value{
The dataframe with an added "Unique" column ("1" for unique, "0" for shared)
}
\description{
Add a "Unique" column to the dataframe.
Main function that combines matrix construction and uniqueness determination.
}
