% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_peptides.R
\name{ApplyUniquePeptidesFilter}
\alias{ApplyUniquePeptidesFilter}
\title{ApplyUniquePeptidesFilter}
\usage{
ApplyUniquePeptidesFilter(df, apply_filter = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{apply_filter}{If TRUE, filter non-unique peptides; if FALSE, only add the Unique column}
}
\value{
Modified dataframe
}
\description{
Simplified compatibility function for easy integration into app.R.
This function is a wrapper around ProcessUniquePeptides with sensible defaults.
}
