% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_peptides.R
\name{ProcessUniquePeptides}
\alias{ProcessUniquePeptides}
\title{ProcessUniquePeptides}
\usage{
ProcessUniquePeptides(
  df,
  save_debug = TRUE,
  debug_path = file.path(getwd(), "data_with_unique_col.tsv"),
  filter_unique = TRUE,
  peptide_col = "Precursor.Id",
  protein_col = "Protein.Group"
)
}
\arguments{
\item{df}{Input dataframe}

\item{save_debug}{If TRUE, save the file with Unique column}

\item{debug_path}{Path to the debug file}

\item{filter_unique}{If TRUE, keep only rows with Unique="1"}

\item{peptide_col}{Name of the peptide/precursor column}

\item{protein_col}{Name of the protein column}
}
\value{
The dataframe (filtered or not depending on filter_unique)
}
\description{
Main function for the complete workflow.
This function combines all steps:
1. Add Unique column
2. Save debug TSV file
3. Filter to keep only unique peptides if requested
}
