% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_peptides.R
\name{SaveDataWithUnique}
\alias{SaveDataWithUnique}
\title{SaveDataWithUnique}
\usage{
SaveDataWithUnique(df, output_path)
}
\arguments{
\item{df}{Dataframe to save}

\item{output_path}{Path to the output file}
}
\description{
Save the dataframe with Unique column as TSV file.
}
